+++ demo-video/remotion/src/Root.tsx
@@ -1,6 +1,7 @@
 import "./index.css";
 import { BuildWeekComposition } from "./Composition";
 import { DeepSeek064LaunchComposition } from "./DeepSeek064Launch";
+import { DeepSeekProductFilmComposition } from "./DeepSeekProductFilm";
 import { DeepSeekReleaseCompositions } from "./DeepSeekRelease";
 
 export const RemotionRoot: React.FC = () => {
@@ -9,6 +10,7 @@ export const RemotionRoot: React.FC = () => {
       <BuildWeekComposition />
       <DeepSeekReleaseCompositions />
       <DeepSeek064LaunchComposition />
+      <DeepSeekProductFilmComposition />
     </>
   );
 };